pyproject.toml
[project]
name = "procrastinate-jobs"
version = "0.5.0"
description = "Understand procrastinate python library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "dramatiq>=1.18.0",
    "dramatiq-pg>=0.12.0",
    "procrastinate>=3.5.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tabulate>=0.9.0",
    "typer>=0.17.4",
]
